begin refactor
Tests
