+++ src/msc/mastercore.cpp
