+++ src/msc/mastercore.h
