+++ src/msc/mastercore_convert.cpp
