+++ src/msc/mastercore_convert.h
