+++ src/msc/mastercore_errors.h
