+++ src/msc/mastercore_parse_string.cpp
