+++ src/msc/mastercore_parse_string.h
