+++ src/msc/mastercore_sp.cpp
