+++ src/msc/mastercore_sp.h
