+++ src/msc/mastercore_version.cpp
