+++ src/msc/mastercore_version.h
