+++ src/test/mastercore_swapbyteorder_tests.cpp
@@ -1,4 +1,4 @@
-#include "mastercore_convert.h"
+#include "msc/mastercore_convert.h"
 
 #include <stdint.h>
